Effects.FileSystem.PathReader

Description

Provides the MonadPathReader effect.

Since: 0.1

Effect

class Monad m => MonadPathReader m where

Represents file-system reader effects.

Since: 0.1

Methods

listDirectory :: HasCallStack => OsPath -> m [OsPath]

Lifted listDirectory.

Since: 0.1

getDirectoryContents :: HasCallStack => OsPath -> m [OsPath]

Lifted getDirectoryContents.

Since: 0.1

getCurrentDirectory :: HasCallStack => m OsPath

Lifted getCurrentDirectory.

Since: 0.1

getHomeDirectory :: HasCallStack => m OsPath

Lifted getHomeDirectory.

Since: 0.1

getXdgDirectory :: HasCallStack => XdgDirectory -> OsPath -> m OsPath

Lifted getXdgDirectory.

Since: 0.1

getXdgDirectoryList :: HasCallStack => XdgDirectoryList -> m [OsPath]

Lifted getXdgDirectoryList.

Since: 0.1

getAppUserDataDirectory :: HasCallStack => OsPath -> m OsPath

Lifted getAppUserDataDirectory.

Since: 0.1

getUserDocumentsDirectory :: HasCallStack => m OsPath

Lifted getUserDocumentsDirectory.

Since: 0.1

getTemporaryDirectory :: HasCallStack => m OsPath

Lifted getTemporaryDirectory.

Since: 0.1

getFileSize :: HasCallStack => OsPath -> m Integer

Lifted getFileSize.

Since: 0.1

canonicalizePath :: HasCallStack => OsPath -> m OsPath

Lifted canonicalizePath.

Since: 0.1

makeAbsolute :: HasCallStack => OsPath -> m OsPath

Lifted makeAbsolute.

Since: 0.1

makeRelativeToCurrentDirectory :: HasCallStack => OsPath -> m OsPath

Lifted makeRelativeToCurrentDirectory.

Since: 0.1

doesPathExist :: HasCallStack => OsPath -> m Bool

Lifted doesPathExist.

Since: 0.1

doesFileExist :: HasCallStack => OsPath -> m Bool

Lifted doesFileExist.

Since: 0.1

doesDirectoryExist :: HasCallStack => OsPath -> m Bool

Lifted doesDirectoryExist.

Since: 0.1

findExecutable :: HasCallStack => OsPath -> m (Maybe OsPath)

Lifted findExecutable.

Since: 0.1

findExecutables :: HasCallStack => OsPath -> m [OsPath]

Lifted findExecutables.

Since: 0.1

findExecutablesInDirectories :: HasCallStack => [OsPath] -> OsPath -> m [OsPath]

Lifted findExecutablesInDirectories.

Since: 0.1

findFileWith :: HasCallStack => (OsPath -> m Bool) -> [OsPath] -> OsPath -> m (Maybe OsPath)

Lifted findFileWith.

Since: 0.1

findFilesWith :: HasCallStack => (OsPath -> m Bool) -> [OsPath] -> OsPath -> m [OsPath]

Lifted findFilesWith.

Since: 0.1

pathIsSymbolicLink :: HasCallStack => OsPath -> m Bool

Lifted pathIsSymbolicLink.

Since: 0.1

getSymbolicLinkTarget :: HasCallStack => OsPath -> m OsPath

Lifted getSymbolicLinkTarget.

Since: 0.1

getPermissions :: HasCallStack => OsPath -> m Permissions

Lifted getPermissions.

Since: 0.1

getAccessTime :: HasCallStack => OsPath -> m UTCTime

Lifted getAccessTime.

Since: 0.1

getModificationTime :: HasCallStack => OsPath -> m UTCTime

Lifted getModificationTime.

Since: 0.1

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

Functions

findFile :: (HasCallStack, MonadPathReader m) => [OsPath] -> OsPath -> m (Maybe OsPath)

Search through the given list of directories for the given file.

The behavior is equivalent to findFileWith, returning only the first occurrence. Details can be found in the documentation of findFileWith.

Since: 0.1

findFiles :: (HasCallStack, MonadPathReader m) => [OsPath] -> OsPath -> m [OsPath]

Search through the given list of directories for the given file and returns all paths where the given file exists.

The behavior is equivalent to findFilesWith. Details can be found in the documentation of findFilesWith.

Since: 0.1

XDG Utils

getXdgData :: (HasCallStack, MonadPathReader m) => OsPath -> m OsPath

Retrieves the XDG data directory e.g. ~/.local/share.

Since: 0.1

getXdgConfig :: (HasCallStack, MonadPathReader m) => OsPath -> m OsPath

Retrieves the XDG config directory e.g. ~/.config.

Since: 0.1

getXdgCache :: (HasCallStack, MonadPathReader m) => OsPath -> m OsPath

Retrieves the XDG cache directory e.g. ~/.cache.

Since: 0.1

getXdgState :: (HasCallStack, MonadPathReader m) => OsPath -> m OsPath

Retrieves the XDG state directory e.g. ~/.local/state.

Since: 0.1

Path Types

data PathType

Path type.

Since: effects-unix-compat-0.1

Constructors

PathTypeFile
Since: effects-unix-compat-0.1

PathTypeDirectory
Since: effects-unix-compat-0.1

PathTypeSymbolicLink
Since: effects-unix-compat-0.1

PathTypeOther
Since: effects-unix-compat-0.1

Functions

displayPathType :: IsString a => PathType -> a

String representation of PathType.

Since: effects-unix-compat-0.1

getPathType :: (HasCallStack, MonadCatch m, MonadPathReader m) => OsPath -> m PathType

Returns the type for a given path without following symlinks. Throws IOException if the path does not exist or the type cannot be detected.

For a faster version in terms of PosixCompat, see effects-unix-compat.

Since: 0.1

isPathType

Arguments

:: (HasCallStack, MonadCatch m, MonadPathReader m)

=> PathType
Expected path type.

-> OsPath

-> m Bool

Checks that the path type matches the expectation. Throws IOException if the path does not exist or the type cannot be detected.

For a faster version in terms of PosixCompat, see effects-unix-compat.

Since: 0.1

throwIfWrongPathType

Arguments

:: (HasCallStack, MonadCatch m, MonadPathReader m)

=> String
The location for the thrown exception (e.g. function name)

-> PathType
Expected path type

-> OsPath
Path

-> m ()

Throws IOException if the path does not exist or the expected path type does not match actual.

For a faster version in terms of PosixCompat, see effects-unix-compat.

Since: 0.1

Optics

_PathTypeFile :: Prism' PathType ()

Since: effects-unix-compat-0.1

_PathTypeDirectory :: Prism' PathType ()

Since: effects-unix-compat-0.1

_PathTypeSymbolicLink :: Prism' PathType ()

Since: effects-unix-compat-0.1

_PathTypeOther :: Prism' PathType ()

Since: effects-unix-compat-0.1

Misc

listDirectoryRecursive

Arguments

:: forall m. (HasCallStack, MonadPathReader m)

=> OsPath
Root path.

-> m ([OsPath], [OsPath])
(files, directories)

Retrieves the recursive directory contents; splits the sub folders and directories apart.

Since: 0.1

listDirectoryRecursiveSymbolicLink

Arguments

:: forall m. (HasCallStack, MonadCatch m, MonadPathReader m)

=> OsPath
Root path.

-> m ([OsPath], [OsPath], [OsPath])
(files, directories, symbolic links)

Like listDirectoryRecursive except symbolic links are not traversed i.e. they are returned separately.

Since: 0.1

doesSymbolicLinkExist :: (HasCallStack, MonadCatch m, MonadPathReader m) => OsPath -> m Bool

Returns true if the path is a symbolic link. Does not traverse the link.

Since: 0.1

pathIsSymbolicDirectoryLink :: (HasCallStack, MonadPathReader m) => OsPath -> m Bool

Returns true if p is a symbolic link and it points to an extant directory. Throws an exception if the path is not a symbolic link or the target does not exist.

This function and pathIsSymbolicFileLink are intended to distinguish file and directory links on Windows. This matters for knowing when to use:

createFileLink vs. createDirectoryLink

removeFile vs. removeDirectoryLink

Suppose we want to copy an arbitrary path p. We first determine that p is a symlink via doesSymbolicLinkExist. If pathIsSymbolicDirectoryLink returns true then we know we should use Effects.FileSystem.PathWriter's createDirectoryLink. Otherwise we can fall back to createFileLink.

Because this relies on the symlink's target, this is best effort, and it is possible pathIsSymbolicDirectoryLink and pathIsSymbolicFileLink both return false.

Note that Posix makes no distinction between file and directory symbolic links. Thus if your system only has to work on Posix, you probably don't need this function.

Since: 0.1

pathIsSymbolicFileLink :: (HasCallStack, MonadPathReader m) => OsPath -> m Bool

Like pathIsSymbolicDirectoryLink but for files.

Since: 0.1

Re-exports

data XdgDirectory

Special directories for storing user-specific application data, configuration, and cache files, as specified by the XDG Base Directory Specification.

Note: On Windows, XdgData and XdgConfig usually map to the same directory.

Since: directory-1.2.3.0

Constructors

XdgData
For data files (e.g. images). It uses the XDG_DATA_HOME environment variable. On non-Windows systems, the default is ~/.local/share. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /usr/share.

XdgConfig
For configuration files. It uses the XDG_CONFIG_HOME environment variable. On non-Windows systems, the default is ~/.config. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /etc.

XdgCache
For non-essential files (e.g. cache). It uses the XDG_CACHE_HOME environment variable. On non-Windows systems, the default is ~/.cache. On Windows, the default is %LOCALAPPDATA% (e.g. C:/Users/<user>/AppData/Local). Can be considered as the user-specific equivalent of /var/cache.

XdgState
For data that should persist between (application) restarts, but that is not important or portable enough to the user that it should be stored in XdgData. It uses the XDG_STATE_HOME environment variable. On non-Windows sytems, the default is ~/.local/state. On Windows, the default is %LOCALAPPDATA% (e.g. C:/Users/<user>/AppData/Local).

Since: directory-1.3.7.0

data XdgDirectoryList

Search paths for various application data, as specified by the XDG Base Directory Specification.

The list of paths is split using searchPathSeparator, which on Windows is a semicolon.

Note: On Windows, XdgDataDirs and XdgConfigDirs usually yield the same result.

Since: directory-1.3.2.0

Constructors

XdgDataDirs
For data files (e.g. images). It uses the XDG_DATA_DIRS environment variable. On non-Windows systems, the default is /usr/local/share/ and /usr/share/. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

XdgConfigDirs
For configuration files. It uses the XDG_CONFIG_DIRS environment variable. On non-Windows systems, the default is /etc/xdg. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

data Permissions

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)