Effects.FileSystem.Utils

Description

Provides filesystem utilities.

Since: 0.1

File paths

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

Encoding

Total

encodeFpToOs :: FilePath -> Either EncodingException OsPath

Encodes a FilePath to an OsPath. This is a pure version of filepath's encodeUtf that returns the EncodingException in the event of an error.

Since: 0.1

encodeFpToOsLenient :: FilePath -> OsPath

Total conversion from FilePath to OsPath, replacing encode failures with the closest visual match.

Since: 0.1

Partial

encodeFpToOsThrowM :: (HasCallStack, MonadThrow m) => FilePath -> m OsPath

encodeFpToOs that throws EncodingException.

Since: 0.1

encodeFpToOsFail :: MonadFail m => FilePath -> m OsPath

encodeFpToOsThrowM with MonadFail.

Since: 0.1

unsafeEncodeFpToOs :: HasCallStack => FilePath -> OsPath

Unsafely converts a FilePath to OsPath falling back to error.

WARNING: Partial

Since: 0.1

Encoding + Validation

Total

osp :: QuasiQuoter

QuasiQuote an OsPath. This accepts Unicode characters and encodes as UTF-8 on unix and UTF-16LE on windows. Runs isValid on the input.

encodeFpToValidOs :: FilePath -> Either EncodingException OsPath

encodeFpToOs that also checks isValid i.e. encodeFpToOs only succeeds if the FilePath can be encoded and passes expected invariants.

Since: 0.1

encodeFpToValidOsLenient :: FilePath -> OsPath

Total conversion from FilePath to OsPath, replacing encode failures with the closest visual match. If the result is not valid, makes it valid.

Since: 0.1

Partial

encodeFpToValidOsThrowM :: (HasCallStack, MonadThrow m) => FilePath -> m OsPath

encodeFpToValidOs that throws EncodingException.

Since: 0.1

encodeFpToValidOsFail :: MonadFail m => FilePath -> m OsPath

encodeFpToValidOs with MonadFail.

Since: 0.1

unsafeEncodeFpToValidOs :: HasCallStack => FilePath -> OsPath

Unsafely converts a FilePath to OsPath falling back to error.

WARNING: Partial

Since: 0.1

Decoding

Total

decodeOsToFp :: OsPath -> Either EncodingException FilePath

Decodes an OsPath to a FilePath. This is a pure version of filepath's decodeUtf.

Since: 0.1

decodeOsToFpLenient :: OsPath -> FilePath

Total conversion from OsPath to FilePath, replacing decode failures with the closest visual match.

Since: 0.1

decodeOsToFpDisplayEx :: OsPath -> String

Total conversion from OsPath to String. If decoding fails, displays the exception.

Since: 0.1

decodeOsToFpShow :: OsPath -> String

Total conversion from OsPath to String. If decoding fails, falls back to its Show instance.

Since: 0.1

Partial

decodeOsToFpThrowM :: (HasCallStack, MonadThrow m) => OsPath -> m FilePath

decodeOsToFp that throws EncodingException.

Since: 0.1

decodeOsToFpFail :: MonadFail m => OsPath -> m FilePath

decodeOsToFp with MonadFail.

Since: 0.1

unsafeDecodeOsToFp :: HasCallStack => OsPath -> FilePath

Unsafely converts an OsPath to a FilePath falling back to error.

WARNING: Partial

Since: 0.1

Functions

(</>) :: OsPath -> OsPath -> OsPath

Combine two paths with a path separator. If the second path starts with a path separator or a drive letter, then it returns the second. The intention is that readFile (dir </> file) will access the same file as setCurrentDirectory dir; readFile file.

Posix:   "/directory" </> "file.ext" == "/directory/file.ext"
Windows: "/directory" </> "file.ext" == "/directory\\file.ext"
         "directory" </> "/file.ext" == "/file.ext"
Valid x => (takeDirectory x </> takeFileName x) `equalFilePath` x

Combined:

Posix:   "/" </> "test" == "/test"
Posix:   "home" </> "bob" == "home/bob"
Posix:   "x:" </> "foo" == "x:/foo"
Windows: "C:\\foo" </> "bar" == "C:\\foo\\bar"
Windows: "home" </> "bob" == "home\\bob"

Not combined:

Posix:   "home" </> "/bob" == "/bob"
Windows: "home" </> "C:\\bob" == "C:\\bob"

Not combined (tricky):

On Windows, if a filepath starts with a single slash, it is relative to the root of the current drive. In [1], this is (confusingly) referred to as an absolute path. The current behavior of </> is to never combine these forms.

Windows: "home" </> "/bob" == "/bob"
Windows: "home" </> "\\bob" == "\\bob"
Windows: "C:\\home" </> "\\bob" == "\\bob"

On Windows, from [1]: "If a file name begins with only a disk designator but not the backslash after the colon, it is interpreted as a relative path to the current directory on the drive with the specified letter." The current behavior of </> is to never combine these forms.

Windows: "D:\\foo" </> "C:bar" == "C:bar"
Windows: "C:\\foo" </> "C:bar" == "C:bar"

(<.>) :: OsPath -> OsString -> OsPath

Add an extension, even if there is already one there, equivalent to addExtension.

"/directory/path" <.> "ext" == "/directory/path.ext"
"/directory/path" <.> ".ext" == "/directory/path.ext"

(-<.>) :: OsPath -> OsString -> OsPath

Remove the current extension and add another, equivalent to replaceExtension.

"/directory/path.txt" -<.> "ext" == "/directory/path.ext"
"/directory/path.txt" -<.> ".ext" == "/directory/path.ext"
"foo.o" -<.> "c" == "foo.c"

Legacy

(</>!) :: HasCallStack => OsPath -> FilePath -> OsPath infixl 9

Unsafely combines an OsPath and a FilePath via (/) with unsafeEncodeFpToOs.

WARNING: Partial

Since: 0.1

(!</>) :: HasCallStack => FilePath -> OsPath -> OsPath infixl 9

Unsafely combines a FilePath and an OsPath via (/) with unsafeEncodeFpToOs.

WARNING: Partial

Since: 0.1

combineFilePaths :: FilePath -> FilePath -> FilePath

Legacy alias for FilePaths' / operator. Exists because the / exported here is (</>) :: OsPath -> OsPath -> OsPath.

Since: 0.1

IO actions

readBinaryFileIO :: OsPath -> IO ByteString

Since: 0.1

writeBinaryFileIO :: OsPath -> ByteString -> IO ()

Since: 0.1

appendBinaryFileIO :: OsPath -> ByteString -> IO ()

Since: 0.1

openBinaryFileIO :: OsPath -> IOMode -> IO Handle

Since: 0.1

withBinaryFileIO :: OsPath -> IOMode -> (Handle -> IO a) -> IO a

Since: 0.1

Decoding UTF-8

Total

decodeUtf8 :: ByteString -> Either UnicodeException Text

Decodes a ByteString to UTF-8.

Since: 0.1

decodeUtf8Lenient :: ByteString -> Text

Leniently decodes a ByteString to UTF-8.

Since: 0.1

Partial

decodeUtf8ThrowM :: (HasCallStack, MonadThrow m) => ByteString -> m Text

Decodes a ByteString to UTF-8. Can throw UnicodeException.

Since: 0.1

decodeUtf8Fail :: MonadFail m => ByteString -> m Text

Decodes a ByteString to UTF-8.

Since: 0.1

unsafeDecodeUtf8 :: HasCallStack => ByteString -> Text

Decodes a ByteString to UTF-8.

WARNING: Partial

Since: 0.1

Encoding UTF-8

encodeUtf8 :: Text -> ByteString

Encode text using UTF-8 encoding.

Misc

throwPathIOError

Arguments

:: (HasCallStack, MonadThrow m)

=> OsPath
Path upon which the IO operation failed.

-> String
String location (e.g. function name).

-> IOErrorType
Type of exception.

-> String
Description.

-> m a

Helper for throwing IOException.

Since: 0.1

(>.>) :: (a -> b) -> (b -> c) -> a -> c infixl 9

Flipped (.).

Since: 0.1